Update Lt-Core files

+++ src/lib/Lt-Core.cc
@@ -31,7 +31,7 @@
 
 #include "../../include/Lt-Core.hh"
 
-void Lt_Core(Display *Dsp, int _CONS){
+void Lt_Core(Display *Dsp, short int _CONS){
     // Software service core function
     D_Lt_Core _D_Core_;
